requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.14
numpy==1.26.4
pytesseract==0.3.13
Pillow==10.4.0